// src/app/modules/configuration/finance/models/fee-type-tranches-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * FeeTypeTrancheView
 */
export interface FeeTypeTranchesView { 
    /**
     * Gets or Sets FeeTypeId
     */
    feeTypeId: number;
    /**
     * Gets or Sets Name
     */
    name: string;
    /**
     * Gets or Sets Code
     */
    code?: string;
    /**
     * Gets or Sets Description
     */
    description?: string;
    id: string;
}